// src/Types/Movie.ts
export interface MovieTitle {
  title: string;
  year: string;
  imdbId: string;
  poster: string;
}

export interface MovieDetails {
  title: string;
  year: string;
  runtime: string;
  genre: string;
  director: string;
  writer: string;
  actors: string;
  plot: string;
  language: string;
  poster: string;
}